Improved loading state UI on Newsletters tab
ref https://linear.app/ghost/issue/PROD-2007/update-loading-states

- Loading states were inconsistent across the various tabs of Analytics. This PR clears up loading state for Analytics / Newsletters
