+++ apps/stats/src/views/Stats/Growth/index.ts
@@ -1,0 +1,1 @@
+export {default} from './Growth';
