+++ apps/stats/src/views/Stats/Newsletters/index.ts
@@ -1,0 +1,1 @@
+export {default} from './Newsletters';
